+++ app.js
@@ -4,7 +4,7 @@ const cors = require("cors");
 
 const authRouter = require("./routes/api/auth");
 const usersRouter = require("./routes/api/users");
-const contactsRouter = require("./routes/api/contacts");
+const transactiosRouter = require("./routes/api/transactions");
 
 const app = express();
 
@@ -17,7 +17,7 @@ app.use(express.static("public"));
 
 app.use("/api/auth", authRouter);
 app.use("/api/users", usersRouter);
-app.use("/api/contacts", contactsRouter);
+app.use("/api/transactions", transactiosRouter);
 
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
