--- helpers/index.js
@@ -1,5 +1,0 @@
-const sendEmail = require("./sendEmail");
-
-module.exports = {
-  sendEmail,
-};
