--- models/contact.js
@@ -1,33 +1,0 @@
-const { Schema, model } = require("mongoose");
-const Joi = require("joi");
-
-const contactSchema = Schema({
-  name: {
-    type: String,
-    required: [true, "Set name for contact"],
-  },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-});
-
-const joiSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-
-const Contact = model("contact", contactSchema);
-
-module.export = {
-  Contact,
-  joiSchema,
-};
