+++ models/index.js
@@ -1,7 +1,7 @@
-const { Contact } = require("./contact");
+const { Transaction } = require("./transaction");
 const { User } = require("./user");
 
 module.exports = {
-  Contact,
+  Transaction,
   User,
 };
